Updated navX_ftc to use the 2016 beta 2.0 libraries.  Also updated OpModes to be compliant with 2016 FTC beta 2.0 library conventions.
